pyproject.toml
[project]
name = "videogame-semantic-search"
version = "0.1.0"
description = "Semantic Web project: video game knowledge graph with NL search powered by SPARQL + LLM"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rdflib>=7.0.0",
    "SPARQLWrapper>=2.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "openai>=1.12.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "isort>=5.13.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.black]
target-version = ["py312"]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
ignore = ["E741"]
